requirements.txt
google-cloud-aiplatform
pandas
numpy
scikit-learn
joblib